:recycle: ユーザー管理機能を分離し、ユーザー登録と存在確認のメソッドを実装
